+++ CHANGES.md
@@ -97,6 +97,7 @@ Fixed:
   could lead to incorrectly randomized playlists (#2507, #2500).
 - Fixed srt output in listener mode to allow more than one listener at a time and
   prevent listening socket from being re-created on listener disconnection (#2556)
+- Fixed race condition when switching `input.ffmpeg`-based urls (#2956)
 
 ---